Rename package 'service' -> 'services' in the mail module
