+++ modules/core/server-scala/src/main/scala/modules/contact/guice/modules/ContactModule.scala
@@ -3,8 +3,8 @@ package modules.contact.guice.modules
 import com.github.jurajburian.mailer.Message
 import modules.contact.graphql.resolvers.{ContactResolver, ContactResolverImpl}
 import modules.mail.models.MailPayload
-import modules.mail.service.MailService
-import modules.mail.service.ethereal.EtherealMailService
+import modules.mail.services.MailService
+import modules.mail.services.ethereal.EtherealMailService
 import net.codingwell.scalaguice.ScalaModule
 
 class ContactModule extends ScalaModule {
